requirements.txt
meilisearch
pandas
flask
jinja2
python-slugify